backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-sqlalchemy==2.5.1
python-dotenv==0.17.1
pyodbc==4.0.32
